requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pydantic
psycopg2-binary
razorpay